requirement.txt
psycopg2-binary==2.8.5
aiopg
boto3